requirements.txt
Tkinter
customtkinter
pymavlink
pyserial
numpy
Pillow
matplotlib
opencv-python